Move binding code out of Migration folder
Fixes #4409
